+++ base/shell/explorer/precomp.h
@@ -339,6 +339,7 @@ HRESULT CTrayClockWnd_CreateInstance(HWND hwndParent, REFIID riid, void **ppv);
 /* TrayNotifyWnd */
 #define TNWM_GETMINIMUMSIZE (WM_USER + 0x100)
 #define TNWM_CHANGETRAYPOS  (WM_USER + 0x104)
+#define TNWM_GETSHOWDESKTOPBUTTON (WM_USER + 0x7601)
 
 #define NTNWM_REALIGN   (0x1)
 
@@ -348,7 +349,6 @@ HRESULT CTrayNotifyWnd_CreateInstance(HWND hwndParent, REFIID riid, void **ppv);
 HRESULT CSysPagerWnd_CreateInstance(HWND hwndParent, REFIID riid, void **ppv);
 
 #include "CTrayShowDesktopButton.h"
-#include "trayntfy.h"
 
 /*
  * taskswnd.c
